use int instead of float in grid_multi offset keys

+++ src/spaces/grid_multi.jl
@@ -5,9 +5,9 @@ struct GridSpace{D,P} <: AbstractGridSpace{D,P}
     stored_ids::Array{Vector{Int},D}
     metric::Symbol
     offsets_at_radius::Dict{Int,Vector{NTuple{D,Int}}}
-    offsets_within_radius::Dict{Float64,Vector{NTuple{D,Int}}}
-    offsets_within_radius_no_0::Dict{Float64,Vector{NTuple{D,Int}}}
-    indices_within_radius_tuple::Dict{NTuple{D,Float64},Vector{NTuple{D,Int}}}
+    offsets_within_radius::Dict{Int,Vector{NTuple{D,Int}}}
+    offsets_within_radius_no_0::Dict{Int,Vector{NTuple{D,Int}}}
+    indices_within_radius_tuple::Dict{NTuple{D,Int},Vector{NTuple{D,Int}}}
 end
 
 """
@@ -74,9 +74,9 @@ function GridSpace(
         stored_ids,
         metric,
         Dict{Int,Vector{NTuple{D,Int}}}(),
-        Dict{Float64,Vector{NTuple{D,Int}}}(),
-        Dict{Float64,Vector{NTuple{D,Int}}}(),
-        Dict{NTuple{D,Float64},Vector{NTuple{D,Int}}}(),
+        Dict{Int,Vector{NTuple{D,Int}}}(),
+        Dict{Int,Vector{NTuple{D,Int}}}(),
+        Dict{NTuple{D,Int},Vector{NTuple{D,Int}}}(),
     )
 end
